requirements.txt
aiohttp==3.8.4
Brotli==1.0.9
openai==0.27.6
python-dotenv==1.0.0
Requests==2.31.0
wordfilter==0.2.7
